More peer review suggestions.
Also, fixed drag-by-mouse scrolling, which was rounding off subpixel
components and losing accuracy when dragging slowly.

+++ source/PreferencesPanel.h
@@ -102,8 +102,8 @@ private:
 
 	std::unique_ptr<RenderBuffer> pluginListClip;
 	std::unique_ptr<RenderBuffer> pluginDescriptionBuffer;
-	ScrollVar<int> pluginListScroll;
-	ScrollVar<int> pluginDescriptionScroll;
+	ScrollVar<double> pluginListScroll;
+	ScrollVar<double> pluginDescriptionScroll;
 	int pluginListHeight = 0;
 };
